requirements.txt
amqp==2.6.1
asgiref==3.6.0
backports.zoneinfo==0.2.1
billiard==3.6.3.0
celery==4.4.7
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.5.0
django-environ==0.9.0
django-filter==2.3.0
django-phone-field==1.8.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==4.5.0
docopt==0.6.2
filelock==3.8.0
Flask==2.2.2
flower==0.9.5
gunicorn==20.0.4
humanize==2.6.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==4.6.11
Markdown==3.4.1
MarkupSafe==2.1.1
num2words==0.5.12
pdf2image==1.16.0
Pillow==8.3.0
platformdirs==2.5.3
prometheus-client==0.8.0
psycopg2==2.9.5
psycopg2-binary==2.8.6
Pygments==2.14.0
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
sentry-sdk==1.5.1
sqlparse==0.3.1
tabulate==0.9.0
tornado==6.0.4
tzdata==2022.6
urllib3==1.26.13
vine==1.3.0
virtualenv==20.16.6
Werkzeug==2.2.2
zipp==3.10.0
